refactoring paterns
