+++ Anvov/Decorator/API/ConcertInterface.php
@@ -7,12 +7,7 @@
  * @author Andriy Vovk <andriy.vovk@smile-ukraine.com>
  * @copyright 2019 Smile
  */
-/**
- * Interface PhoneInterface
- *
- * Made to describe functions of regular phone
- * @package school5\StaticFactory\API
- */
+
 interface ConcertInterface
 {
     /**
